Velocity extrapolation vof (#771)
Description of the problem
The way the velocity field in the VOF equation was calculated implied the use of an explicit Euler scheme for that component. For the wave oscillation problem (sloshing in a rectangular tank) this lead to very poor results
Description of the solution
Add a time-extrapolation of the velocity field in the VOF equations
Add a new example for sloshing in a rectangular tank
How Has This Been Tested?
A new unit test has been added for the time extrapolation. It tests various order of extrapolation
The new benchmark test really tests this deeply
Comment
This was an important piece that was missing. Happy to see things are good now.



Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>

Former-commit-id: 3238e8bb1608304ac491fefb33a7b547be271349