Execute particle displacement only when sort particles in cells (#786)
Description of the problem
When particles cross PBC, their location is updated but not the cell in which they are. No issues was encountered in the pass, bugs seem to happen when diameter is very small.

Description of the solution
Move the execution of the particle displacement at the same step as the sorting of particles in cells. This means the particles won't have their location updated as soon as the position > the location of the periodic wall. The particle displacement is already executed as is for CFD-DEM.

How Has This Been Tested?
Tested with hopper case and tests are good

Former-commit-id: 0f35d9167a63cdfa9a6daecadcb8cb92e10de49d