Add isothermal ideal gas density model (#785)
Description of the problem
When experimenting with the dam break with an obstacle example, it was noticed that the software performed poorly in mimicking the wave's natural shape after the impact. It was assumed that this is due to the lack of considering the air's natural compressibility.

Description of the solution
In an attempt to solve this issue, an isothermal compressible fluid dynamics model will be implemented. As a first step, an isothermal ideal gas density model is implemented.

How Has This Been Tested?
A unit test for the isothermal ideal gas density model was added.

 tests/core/density_isothermal_ideal_gas.cc

Documentation
The physical properties page in the documentation was updated by adding the isothermal ideal gas density model.

 doc/source/parameters/cfd/physical_properties.rst

Future changes
Currently, the isothermal ideal gas model can only be used in conjunction with the incompressible Navier-Stokes equations. However, it is meant to be used with the isothermal formulation of compressible Navier-Stokes equations to account for weakly compressible fluid flows. In a future PR, an isothermal compressible formulation of the equations will be implemented.
Comments
A DataPostprocessor object was implemented for density, and the density field is only outputted when it is not constant.
Other minor corrections were made to unrelated files (mainly typos in the comments).

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Former-commit-id: ca7783e3c8e1397548c3fc7cbcf5d192e6d6a966