Interactive diagrams for the examples in the documentation (#800)
Description of the problem
There was no interactive way of going through the examples in the documentation.

Description of the solution
An interactive diagram/map for each of the categories was created. Now it is easier to find examples according to specific criteria. The title of some examples were modified (e.g., dimensions removed for incompressible flow examples, and only the first capital letter of the titles was kept with exception of names or capital letters needed for acronyms).

Comments
The graphviz package must be installed for the diagrams to compile. The instructions were added in the contributing tab of the documentation and the github workflow for the documentation was also updated accordingly. Two images were added in the diagram for the incompressible flow as a possible way of making it look nicer.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Former-commit-id: c8a5f383feb295e5afb68a199104a0a57c71004e