Ch implementation (#794)
Description of the new materials
New Cahn-Hilliard solver
A new set of equations (Cahn-Hilliard equations) was implemented into Lethe for solving multiphases problems. The aim is not to compete with the VOF solver but rather add new functionnalities, such as a boundary condition on the angle of contact, or naturally diffuse interface.
The solver was added with almost all the usual features for multiphysics, it works both in steady and transitory states. An advective term was also added to take into account the movement of fluid, though it is NOT a coupled Navier-Stokes/Cahn-Hilliard system (this will be for a next PR)

How Has This Been Tested?
A manufactured solution was used as a test, hence the new application_test .
Documentation
A PR is opened in parallel for all the documentation.

Future changes
-Next PR will include a weakly coupled Navier-Stokes and Cahn-Hilliard system. This will allow to have different densities based on the phase field and also take the surface tension effects in the velocity/pressure field.

Co-authored-by: oguevremont <guevremont.o@gmail.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>


Former-commit-id: 5e308b5feb630c52b0080f301bc1a3dfa992e60a