Make usage of Ankerl map more widespread (#812)
Description of the problem
The Ankerl maps were not used optimally everywhere where they could be

Description of the solution
Add this. Also took the opportunity to optimize one part of fine search

How Has This Been Tested?
All unit tests and examples

Comments
This did not change the performance significantly. I will need to profile the code again to identify bottlenecks now.

Co-authored-by: Audrey Collard-Daigneault <71884806+acdaigneault@users.noreply.github.com>
Former-commit-id: 24fa0b0880960a9dfd094ca9ecc4e5162c596e53