Superquadric shapes (#784)
Description of the problem
Superquadric shapes were not implemented.
Superquadric shapes have a surface function that is distorted (gradient isn't systematically of norm 1), which means it can't be used directly to find the gradient and closest surface points.

Description of the solution
An iterative way to find the closest point, and subsequently value and gradient of the distance function, is added.

How Has This Been Tested?
A few new lines were added in core/shape_primitives.cc.
Visual inspections with different parameters were made.

Documentation
The new shape was added to the documentation.

Future changes
For now, the gradient and values found are a bit wrong: gradient is not perpendicular to the surface by construction, so the values can be a bit overestimated. For the purpose of sharp immersed boundary solver this will not be a problem, but a modified objective function to find the closest point could be used to find the true closest points (and so gradient and value). This will be the subject of another PR.

Co-authored-by: PierreLaurentinCS <81017102+PierreLaurentinCS@users.noreply.github.com>
Former-commit-id: 47a3120a19ac0bc0490a1e33b182a772763ab771