Matrix-free Navier-Stokes prototype (#820)
Description of the problem
This PR adds the first matrix free (MF) steady-state Navier-Stokes prototype. It uses SUPG/PSPG stabilization and includes both global coarsening and local smoothing MG approaches as preconditioners for the linear iteration.

Comments
This is the fourth step towards Matrix-Free Lethe. It is the last MF prototype and now all the MF work will be done directly in the Lethe implementation.