Remove dimension of the problem from the executable name (#834)
Description of the problem
The diimension (dim) is a template parameter of the solver classes. The original strategy we had chosen was to have one execxutable per type of solver + dimension. This was becoming unmanageable since we had an increasingly large amount of executable. This PR addresses this by making dimension a parameter of the prm file instead of something in the executable name

Description of the solution
Make the dimension a parameter
Also took the opportunity to deprecate the nitsche_23 solver since nobody uses it and it is not robust at all.

How Has This Been Tested?
All tests have been adapted and they work as before

Documentation
Documentation is updated and finished

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>