Update 2d dam break example (#838)
Description of the problem
The physical properties of the example did not correspond to the ones of air and water, leading to good, but improvable results.
The figure showing the initial state had a mistake (wrong domain length).

Description of the solution
The physical properties of the fluids were changed to ones closer to real air and water at 25 degC and the results were updated.
A new figure for the initial state of the problem was made.

How Has This Been Tested?
By comparing the simulation results with the experimental data of Martin and Moyce (1952).

Documentation
 doc/source/examples/multiphysics/dam-break/dam-break.rst
Comments
For a very low air viscosity, the solver has trouble solving and results are non-physical. Therefore, a pseudo-air with a viscosity of 1.516E-2 is used as fluid 0.

Former-commit-id: f7067cddb2e84825561d682b7c68c2bcf804eaeb