Add matrix free Navier Stokes operator (#836)
Description of the problem
This PR is a follow up on the matrix free Navier-Stokes solver shell created in PR #755. It adds a first working version of the matrix free Navier-Stokes solver in Lethe by implementing a matrix free operator.

Description of the solution
The application was renamed and updated according to the new dimension parameter added in Lethe recently.
For the matrix free operators, a base class was added and a SUPG-PSPG derived class was implemented.
The MF Navier-Stokes solver was updated to use the system operator instead of the system matrix as in the traditional Navier Stokes solver.
How Has This Been Tested?
The application is tested in 2D and 3D by using a MMS. Two element orders are considered in 2D and one in 3D. The following tests were added:

 applications_tests/mf_navier_stokes/mms2d_fe1.prm
 applications_tests/mf_navier_stokes/mms2d_fe2.prm
 applications_tests/mf_navier_stokes/mms3d_fe1.prm
Documentation
No documentation has been added as the solver is not fully functional yet from the user point of view.
Future changes
At the moment, the solver supports only GMRES without preconditioning as linear solver. The next step is to implement the different MG preconditioners. This will lead to the refactoring of linear solvers throughout Lethe.


Former-commit-id: f3fca29fc30ce32766b2fd6de5862cc14f9ae52c