Announce Fluid Dynamics Solver (#848)
Description of the problem
Although all multiphysics are announced before being solved for, this is not the case for the fluid dynamics. This creates unncessary confusion when we have a multiphysics solver

Description of the solution
Announce the fluid dynamics. Regrettably this forced me to update a tremendous amount of test. Some numbers might have changed, but that's because the compiler I am using is not the same than the one that was used when creating the tests.

How Has This Been Tested?
All tests should pass

Comments
This is a first step before having a timer per Physics (for which I will also make a PR today)

Former-commit-id: 5a75e56c63ee8853338fa8369291c13781c15c2b