Find contact detection step interval (#847)
Description of the problem
Currently, we use MPI to communicate at every iteration if the contact detection structure should be rebuilt. This is not necessarily efficient when a large number of core is used.
Description of the solution
I adapted the frequency parameter to also be used in dynamic contact search. In this case, it indicates the frequency at which MPI communications should be done. I have found that increasing the frequency to 2 or 4 can decrease simulation cost by 5-10% in some cases when we use 8 or more cores.
How Has This Been Tested?
All tests still pass
If applicable, describe future changes or features that will be added in subsequent pull requests

Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Former-commit-id: a837646441ed574468c23ca81a3f014dff17af6d