Add dynamic viscosity to rheological models and rename "viscosity" to "kinematic viscosity" (#851)
RDescription of the problem
In multiphase assemblers (NSVOF, ICNSVOF and NSCH) and isothermal compressible NS assembler, dynamic viscosity which is used in the momentum equation was calculated inside the assemblers using the kinematic viscosity. However, with a non-constant density, such as the case of compressible models, the dynamic viscosity defined in the assemblers as dynamic_viscosity = kinematic_viscosity * density was fluctuating when it was not supposed to.
Description of the solution
To solve this issue, a get_dynamic_viscosity(...) and a get_dynamic_viscosity_vector(...) functions were added to the rheological models calculating the non-density-dependent dynamic viscosity.
For the case of a constant dynamic viscosity, the value is only calculated once when the value is set using the set_dynamic_viscosity(...) function. This function is called in source/solvers/physical_properties_manager.cc after getting the rheological model of a fluid.
viscosity that was previously used to refer to a kinematic_viscosity was renamed everywhere except in the names of user-inputted parameters of rheological models.
For all multiphase flow simulations, density and dynamic viscosity fields for each fluid are now outputted. A data postprocessor for dynamic viscosity was added and the one for the kinematic viscosity was renamed.
How Has This Been Tested?
After renaming viscosity to kinematic viscosity, affected .prm files of application tests were updated.
Unit tests testing the dynamic viscosity that were updated are listed below:
 tests/core/carreau_rheology.cc (.output)
 tests/core/newtonian_rheology.cc (.output)
 tests/core/power_law_rheology.cc (.output)
A new unit test was added for the PhaseChange rheological model:
 tests/core/phase_change_rheology.cc (.output)
After changing for the dynamic viscosity in the assemblers, affected application test outputs were updated:
 applications_tests/gls_navier_stokes/cahn_hilliard_navier_stokes_coupling.output
 applications_tests/gls_navier_stokes/gls_vof_1_isothermal_compressible_fluid.output
 applications_tests/gls_navier_stokes/gls_vof_2_isothermal_compressible_fluids.output
 applications_tests/gls_navier_stokes/gls_vof_ht_adapt-dissipation.output
 applications_tests/gls_navier_stokes/gls_vof_peeling.output
 applications_tests/gls_navier_stokes/gls_vof_tanh_filter.output
 applications_tests/gls_navier_stokes/gls_vof_wetting-inverted.output
 applications_tests/gls_navier_stokes/gls_vof_wetting_mesh-adapt.output
 applications_tests/gls_navier_stokes/mms2d_isothermal_compressible_ns_gls.output
 applications_tests/gls_navier_stokes/mms3d_isothermal_compressible_ns_gls.output
Most of VOF examples were runned to check if there were important differences in the results, no deterioration of results was observed.
Documentation
All documentation and examples using an initial kinematic viscosity (parameter of the initial conditions subsection) were updated. Unnecessary default and commented lines in .prm files of examples were removed.
Future changes
Rename parameter names of rheological models to include kinematic when referring to a kinematic viscosity.
Rerun all VOF examples and update results in documentation if necessary.
Comments
During this implementation, it was also noticed that nothing in rheological models indicated if the model was non-Newtonian or not. Therefore, a boolean variable and a function returning that boolean was added to the class. Likewise, for density models, a boolean and function was added to assess if a density model is constant or not. Previously, only a boolean for a whole physics was implemented.
Duplicate allocation of physical properties vectors in HT scratch data was removed.
Density in the calculation of stabilization constant tau was deprecated for dimensional coherence.

Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Former-commit-id: 226e631268e617afe7e135c9a90002a703d94f34