Physic based linear solvers (#853)
Description of the problem
Until now, it was not possible in Lethe to specify different parameters for the linear solver for each of the physics involved in a simulation (e.g., fluid dynamics + VOF, fluid_dynamics + heat transfer).

Description of the solution
Instead of having one struct for the linear solver parameters, a map of structs was created in order to have a different subsection for all of the physics. Therefore, all of the linear solver parameters can be now specified for each physic and they all have the same default values.

How Has This Been Tested?
All the parameter files in the application tests were updated in order to have a subsection in the linear solver parameters subsection that accounted for each of the physics involved. None of the outputs needed to be updated.
The tests/solvers/restart_01.cc test was also updated as it uses the verbosity parameter of the linear solver parameters.

Documentation
The main page describing the linear solver parameters was updated to mention the possibility of specifying parameters for each physic.
The parameter files of the examples were also updated accordingly. Therefore, the examples that explicitly mentioned the linear solver section in the documentation had to be updated as well.

Future changes
A following PR will refactor the linear solver across all applications, which will lead to the addition of a "preconditioner" parameter to the linear solver parameters for each physic.

Former-commit-id: 760bc33d8200f6dca6f98cde055609d0da2d0663