Refactor surface tension gradient parameter (#854)
Description of the problem
After creating a new material interaction class to set interface physical properties, a constant surface tension model was added (Add material interaction class and surface tension model #818). However, to consider Marangoni effet, a linear surface tension model is needed. This model was implicitly implemented in the assemblers.

Description of the solution
With this PR, a temperature-dependent linear surface tension model is added to the surface tension models.
How Has This Been Tested?
The .prm file of the gls_droplet_marangoni_effect application test was updated and a new unit test for the linear surface tension model was added.
 applications_tests/gls_navier_stokes/gls_droplet_marangoni_effect.prm
 tests/core/surface_tension_linear.cc (.output)

Documentation
Documentation on surface tension was updated:
 doc/source/parameters/cfd/physical_properties.rst
 doc/source/parameters/cfd/volume_of_fluid.rst

Future changes
For non-time dependent temperature schemes, it would be useful to implement a position-dependent surface tension model that would avoid the need to solve the heat transfer physics to get the temperature.
Review the formulation of the linear surface tension model, by adding a reference temperature for sigma_0.
Deprecate peeling and wetting.

Comments
VOF parameters documentation was restructured with new subtitles to improve the reading experience.

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: oguevremont <guevremont.o@gmail.com>
Former-commit-id: 9f83f0359427cb323f3e13638c6d72874b8f5385