Dem plane insertion (#828)
Description of the problem
The insertion in DEM can be a real nightmare when using a large number of particles. Particles can bounce back in the insertion zone which can cause "pop-corn' effect. Also, when the whole domain is close to be filled, big overlap between particle is unavoidable when inserting.
This problem makes simulations slower or even can make them unusable went big overlap occurs.

Description of the solution
Plane insertion method uses a group of cells, which are all in the same mathematical plane, to insert new particles.
On every insertion step, we check if those cells are individually empty. If a cell is empty, a new particle will be inserted at it center. This way, no overlap is accounted.
Currently, the insertion stop when a certain number of particle is reach.

How Has This Been Tested?
Insertion in a Hyper_cube with different refinement level have been used.
Different plane (point and normal vector) have be used in that Hyper_cube.
Test and application test have been created.

Documentation
insertion_info.rst -

Future changes
The stop criteria could the the filling of the domain and not the total number of particle inserted.