Finalize shock capture and add switch for interface compressibility (#866)
Description of the problem
The shock capturing approach implemented used an implicit implementation of the added viscosity. This was adequate, but lead to a non-linear system which was very difficult to converge
The interface compressibility term added was wrecking cases with surface tension

Description of the solution
Explicit the added viscosity term
Finalize the formulation to remove the added viscosity in the direction of the velocity vector
Fix the initial condition of the 3D dam break which was very discontinuous
Add a switch to consider (or not) interface compression in the VOF cases. Default value is to disable compressibility effect. The parameter is called compressible and a default value of false.
How Has This Been Tested?
Some tests had to be updated slightly

Comments
This shock capture performs well. I think that's the end of it
Interface compressibility came as a weird surprise. I was not expecting it to have such a dramatic effect inside droplets or within cases with surface tension effects
I will make another PR in the following week to add this to the theory guide!

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>