Deprecate peeling wetting (#869)
Description of the problem
The peeling and wetting mechanism was very ad-hoc and not being used anymore. It was also not robust because of all the changes made to streamline the VOF solver. This is a right time to deprecate it and, if need be in the future or if we find people needing a similar mechanism, we can re-introduce it more robustly (using the compressible equation of state with it for example)

Description of the solution
Deprecate

How Has This Been Tested?
peeling-wetting tests have been removed

Documentation
Peeling-wetting documentation has been removed

Comments
It's cool to deprecate stuff from time to time. I think it's a healthy reflection on the state of the code
I will deprecate the "conservative fluid" stuff next

Co-authored-by: hepap <47506601+hepap@users.noreply.github.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>