Dem insertion list velocity, omega and diameters (#868)
Description of the problem
The list insertion method was only allowing the insertion of static particles. Now, every particle can be inserted with a specific transnational and rotational velocity. The diameter of every particles can also be specify.

Description of the solution
7 new parameters.

A new assign_particle_properties_for_list_insertion function has been created.
How Has This Been Tested?

Test will be added if requested.
The bouncing particle was use to test the different parameters, with success.

Documentation
Insertion section.

Former-commit-id: 97ec2bef115d057529d876c6958c0ec29d456a88