Standardize .prm files comments (#875)
Description of the problem
Comments on .prm files were not uniform;
Some .prm files didn't have #----- commented lines

Description of the solution

Fixed it.
How Has This Been Tested?
N/A

Documentation
Not changed.
Future changes
Update sharp examples (documentation and files).

Comments
At first, I would update the examples only, but I figured this could be done with a small script I wrote so I decided to work on formatting first. The next step is to update the examples.