Bunny drill example (#876)
Description of the problem
We had been thinking about making a bunny drill example for a while. Here it is. It is a bit more light hearted than the other examples.

Description of the solution
The bunny drill is here. It's a chill example. It works quite well actually.

How Has This Been Tested?
Ran the drill

Documentation
Documented the drill

Future changes
Short examples like this one are also a nice way to show features. I like it.

Co-authored-by: Victor Oliveira Ferreira <65647639+voferreira@users.noreply.github.com>
Co-authored-by: Laura Prieto Saavedra <40216050+lpsaavedra@users.noreply.github.com>
Former-commit-id: 572ce02a2868cff46d55bb1d18f786822997a9e5