Rename Lethe applications  (#882)
Description of the problem
The name of the applications in Lethe did not follow any standard format.

Description of the solution
All the applications were renamed according to what was discussed in the group. All the solvers related to fluids start now with lethe-fluid, the DEM solver was renamed as lethe-particles, the coupling of CFD and DEM lethe-fluid-particles, and all rpt applications start now with lethe-rpt.

How Has This Been Tested?
All tests pass since only the name of the folder and the executable was updated.

Documentation
All the examples were updated since the solver used is always mentioned in the ''Features" section. Also, all the commands to run the examples in the "Running the Simulation" section were updated accordingly.

A new file was added to the documentation: doc/source/applications.rst. It adds a table with all the available applications and a brief description of each of them. @blaisb will complete/modify the description of the applications in this PR.

Comments
This PR only changes the name of executable and it does not change the name of any class, .cc or .h file.


Co-authored-by: Bruno Blais <blais.bruno@gmail.com>