Simulation parameters tidy (#885)
Description of the problem
There was an enable parameter in the Source term functionnalities that was poorly implemented. The enable true/false was only used in the Navier-Stokes solver, nowhere else. This led to significant inconsistancies.

Description of the solution
Deprecate the enable parameter
Deprecated the way the Source term works. Now the parameters are a class instead of a pointer to a class, but the functions themselves are pointer. This is significantly cleaner.

How Has This Been Tested?
All tests pass without change.

Documentation
Parameter was removed from documentation and examples
CHANGELOG was modified

Former-commit-id: 7093de01ebcd5b692b02243165e96e14a8313ea7