Add multigrid preconditioners to matrix-free application (#884)
Description of the problem
Until now the matrix-free application in Lethe did not have any preconditioner suitable for the matrix-free approach.

Description of the solution
This PR adds two geometric multigrid approaches: i) local smoothing (LSMG) and ii) global coarsening (GCMG). They were added to the matrix-free application directly. They are used as preconditioners for a GMRES solver; they use the relaxation smoother, and a GMRES coarse-grid solver that uses AMG as preconditioner. The parameters for the coarse-grid solver, the smoother and general MG can be specified via the parameter file.

How Has This Been Tested?
A simple MMS test in 3D and using FE degree 1 was added to test the LSMG preconditioner:

 applications_tests/lethe-fluid-matrix-free/mms3d_fe1_lsmg.prm
The output of several tests unrelated to matrix-free had to be updated. The reason for this is that the triangulation used to have "mesh smoothing", but it was changed to "limit level difference at vertices" as required for the MG preconditioners. This changed slightly the number of cells and therefore the results in several tests.

Documentation
The MG parameters were added to the linear solver section in the documentation:

 doc/source/parameters
Future changes
The next step after merging this PR is to support the different possible BCs and to implement the transient part of the operator.
Comments
There seems to be a bug with the global coarsening algorithm. It runs, but it is extremely slow and the iteration counts suggest that something is not set correctly. A similar test than the one for LSMG can be added, however, at the moment it takes too long.