Dem particle wall tangential overlap (#889)
Description of the problems
The angular velocity of the particle was not taken into account correctly in the contact_relative_velocity calculation. (missing a minus sign)
Particle-Particle and Particle-Wall contact were not using the same vector convention (i -> j)

Description of the solution
The code is now using the same convention for Particle-Particle and Particle-Wall. The normal_contact vector is going from the particle ( i ) to the wall ( j ).
The translational velocity induce by the rotation of the particle at the point of contact is now properly being calculated.
Comments in the code have been added explaining why minus sign and "-=" or "+=" operator are being use at certain places, which makes the calculation of the force way simpler to understand.
An application test have been added. This test rolling_on_plane.prm consisted of a single particle rolling perfectly on a plane without rolling resistance. The tangential_relative_velocity should remain at zero if the contact relation velocity calculation is done correctly.

How Has This Been Tested?
New application test.
Bouncing particle results are the same.
Oblique wall impact collision results are the same. (Not exactly, which is normal)
The influence of the angular velocity of the wall has been tested using the small-rotating-drum example.

Comments
none

Former-commit-id: 1f5fa257856e2c9784d50a109f80584fbdcd78fe