Fix issue with constraints in the VANS solver and update boycott (#895)
Description of the problem
There was a major issue with the void fraction constraints in the VANS and CFD-DEM solver which led to an assertion being thrown in debug mode. The constraints were not reinited with the right size

Description of the solution
Fix this issue by reiniting the constraints correctly
Also fixed an issue with an example while I was there

How Has This Been Tested?
All tests should now pass in debug

Comments
Some part of the VANS solver and the unresolved CFD-DEM solver are still a bit of a mess. I will revisit this slowly but surely within the coming months.