Dem particle wall sliding velocity (#896)
Description of the problem
The calculation of the translational velocity on rotating walls was calculated inadequately #896: The velocity of rotating boundary conditions (e.g. boundaries of the mesh) was calculated inadequately in the case where the normal vector of the wall was aligned with the rotation axis. The whole calculation procedure was slightly messed up and only worked for cylinders.

Description of the solution
This has been fixed and made general by using the vector from the contact point to the center of rotation. A new parameter has been added for this purpose.

How Has This Been Tested?
All test pass and give adequate results
A new application test has been added to test this exact feature

Documentation
Documentation of the point on the axis has been updated

Future changes
I will continue to refactor the entire DEM particle-wall aspect to make it cleaner and more streamlined

Co-authored-by: OGaboriault <102625061+OGaboriault@users.noreply.github.com>