Dem jkr (#899)
Description of the new feature
New force model that takes into account the van der Walls forces. The name of the model is JKR (or SJKR-A)

How Has This Been Tested?
The overlap at the equilibrium has been verify. Two application tests have been created in that regard.

Documentation
Theory, Model Parameters and Lagrangian Physical Proprieties files have been updated.

Future changes
The name of each force model for particle-particle and particle-wall should be the same. Right now, the non-linear modelcorrespond to the hertz_model. The theory guide is not clear on that regard. The documentation doesn't describe the different coulomb limit method that we have.

Comments
NONE