Add viscous and ramp initial conditions for MF (#906)
Description of the problem
The matrix free implementation did not allow for viscous or ramp initial conditions

Description of the solution
Added them.

How Has This Been Tested?
Tested it on a case and it works just fine. Not sure if it needs a dedicated application_test though. @lpsaavedra what's your take on this?

Former-commit-id: d5d6486161e64a3c6a08f53d5eedd6f391669315