Improve output of multigrid information (#914)
Description of the problem
Until now, the only information being printed when the verbosity parameter of the multigrid preconditioners was enabled was the information of the levels. However, it is useful to have more information for debugging or analysis purposes.

Description of the solution
The number of iterations needed by the coarse-grid solver was also added to this output. In addition, one can set the mg verbosity parameter to extra verbose to output a table with the timings required to set up the different components of the multigrid preconditioners. Also, the output of the eigenvalue estimation was fixed and now it reports the corresponding level.

How Has This Been Tested?
The output was tested with different problems on my machine, seems to work as expected.

Documentation
An additional tip was added to the linear solver parameters doc page explaining the difference between verbose and extra verbose for the multigrid preconditioners.

Former-commit-id: 4080590408ef6948b6821d0697f2bbd3057d8211